[IOTMID-48][IOTMID-49][IOTMID-53] Envs update and initial translation
This makes a few changes to the environment, allowing webpack
to properly process the sass defs of bower-retrieved packages
(i.e. font-awesome), hence reducing the number of stylesheet
files to be served on production environment.

This also builds upon the initial commit to provide an initial
"translation" of the angular-1.x based implementation of the
device list page.

Change-Id: I6236e7c82864e5cdab62360c331cb37bac97400c