[IOTMID-51] Fixes materialize select bug
Materialize would not properly forward/register onChange
event callback for the field, which led react's onChange
not to work.

This removes the global jQuery initialization from index.html in
favor of react's own initialization callback, and installs custom
onChange listener for the select field.

Perhaps having a custom materialize-select component would be the
best long-term solution for this (as in react-materialize).

Change-Id: Idd68d792612f6c9a376e69323f9f41006f1cb3ce