[IOTMID-51] Fixes modal close/commit button route issue
This commit removes the bogus href field from the modal JSX, thus
removing any (already existing or future) conflicts with client-side
route definitions.

This also makes a few changes to the navbar and route definitions
allowing better visualization of current page as well as including
the placeholders for future pages.

Change-Id: I12326d6ca87adfc19c20cf141347e43959615e49