[IOTMID-66] Template/Device relation, Icon for templates
This commit changes the utility of the type field for devices
allowing it to reflect the template used when configuring
the given device.

Also, this adds the exhibition of the associated icon on
the templates list as well as code to properly display the
current icon on the template area.

Change-Id: I03c6fd895ec05dd3ca31031cc2f839dc9ae36107