[IOTMID-185] Moves dependencies to <head>
This changes the way JS external dependencies are
declared on the page (moving them from <body> to <head>
as well as adding a ready() guard on when initializing
react on the page.

This is needed due to some cases where a component would
make usage of a dependency before that is properly loaded.

Change-Id: I34187e0550aa9a9c3cd68b53cae2ac9c9fc20aa1