[IOTMID-278] Fixes webpack issue with development setup
This adds a new configuration option to the development
configuration of webpack, allowing devs to redirect
kong to their local-running instances of webpack.

This was needed due to a interface-breaking patch recently
released to the community, by webpack maintainers.

For the minutes and release notes regarding the "security
issue" associated with this:

https://github.com/webpack/webpack-dev-server/releases/tag/v2.4.3
https://medium.com/webpack/webpack-dev-server-middleware-security-issues-1489d950874a



Change-Id: I731e74fcba7391a514b47e8ff4f5c9a88365f301