[IOTMID-322] Fixes display of input-type-options
This changes the rendering of the available options for typed
inputs when defined on a node's form template. More specifically
this changes the DOM insertion point of the option list so it
matches the rule set defined on the new sass/css hierarchy as well
as the positioning (from absolute to fixed) as to encompass the
insertion of the node-red canvas within the higher GUI DOM
structure.

Change-Id: I2e3a3d946d552b326daa20c5cc70f179e70930a6