[IOTMID-440] Makes the action div have fixed width
Should the screen size be smaller, the proportional width set by
sX would, eventually, make the action items be displayed, at the
limit as a column, instead of a row. This makes the width of the
action area fixed, with width of the device name compensating
for the lack of space.

Change-Id: I0d9cad800564dd47ea26fccb3b5bb30d97026125