[IOTMID-573] Fixes positioning of error messages on firefox
On firefox, for some yet unknown reason, the position:absolute
used to allow the horizontal alignment of the error message
ends up overriding the flexbox positioning, causing the div
to be placed on the top of the first relative parent.

This change such positioning to be relative, fixing the error seen
on firefox.

Change-Id: Ic59a617bf3e2635ccc580f3419a4f9eff09ad76c