[IOTMID-449] Adds check for duplicate attr names
This adds a check for duplicate names when editing the set of
attributes for a given device.

Do notice, however, that this is actually a temporary solution,
pending further implementation on the iotagent to allow the
usage of custom keys for fields, as well as the usage of a
system-generated id to tell attributes apart throughout the
system.

Change-Id: Ifffbe93bb48824d4fa6e0de495d3717f8a94f46a