[IOTMID-869]-Integration of the POC with the graphic interface
- This commit implements adjustment in the user screen to allow integration
of the graphic interface with the POC (Integrate PEPKong and KeyPass
into the platform).
- Inclusion of the profile field on the user's screen.

(+)- Correction: The user was required to select a profile,
if he did not select anything the profile field was empty.
The fix was made so that if the user does not select any
profile, by default he is admin.

Change-Id: I6b7b16604d535408c9fe9d32642b5d0e415e598d