[IOTMID-915][IOTMID-768] Fixes user data cache usage
This fixes a bug where creation and updates to users
would invalidate local copy of user data, with a couple of
fields having their values emptied.

Also, this fixes some referencing errors that would occur
when editing active user data (that is, users that are
already configured on the system), making it seem that
data has been saved, when it really was not.

Change-Id: I6582de620e5cf9754533290c355d06cd096f4ea9