[IOTMID-769] Changes error messages for user fields
This changes the error messages for the user form, setting them
to the ones described at their respective jira tickets.

Also, this deprecates HTML5 validation scheme in favor of a
more react-friendly approach.

As a bonus, password field client-side validation can now
account for the semantic differences of editing a user as
opposed to creating one - thus allowing the password field
to be required only on user creation (behavior for user
edition is not to alter the user's password if not set on
the request payload).

Affected jira tickets:

[IOTMID-593]
[IOTMID-592]
[IOTMID-765]
[IOTMID-922]

Change-Id: I8cbc832c7e381cd04186c56fde1bd24eb6a9f979