[IOTMID-758] Fixes type validation regexp, error messages location
This changes the regexp used to validate whether the static value
set to an attribute conforms to the type of the attribute to be
created.

Also, this normalizes the exhibition of error messages, diplaying
them in the context (field) that is in error, in lieu of having a
top error header being displayed in the attribute creation modal.

Change-Id: I52dc6d9fd5ea8992a5aee8e793cbf644723b85b0