[IOTMID-999] Fixes value parsing from STH
The formatting of values from STH is actually denpendent on the
way orion's subscription handler parses them. That may cause numerical
types to be JSON-wise formatted as integers, floats (number) or as
strings.

This fixes a bug where an integer attribute being sent as an integer
(again, JSON-wise) would make the rendering of the device detail
fail.

Change-Id: I842d176c1d3993f2064d3291a191d332bb1ee86a