Fixes device map view.
The number of devices on the map (shown in the "Showing ..." label) is related
to the number of devices that have published geolocation value. If a device is
created but does not yet have a published geolocation value, that device does not
appear on the map, so the number displayed on the label will not change.