[issue#1346] Fixed component that use web socket real time data
The SocketIO client was incorporated as an attribute of the ViewDeviceComponent component,
now it is initialized in the react component constructor, but does not automatically
establish a web socket connection, it is now manual and is only established when the
component is mounted by React. Similarly, the connection is terminated when the component
is unmounted by React. If there are any errors terminating the connection to the
SocketIO Server, reconnection is done manually because a new one-time token must
be generated before reconnecting.